make json output consistent across tap commands
