+++ cli/src/tap/tap_bind.rs
@@ -1,7 +1,4 @@
-use {
-    super::*,
-    nexus_sdk::nexus::tap::{BindAgentSkillParams, BindAgentSkillResult},
-};
+use {super::*, nexus_sdk::nexus::tap::BindAgentSkillParams};
 
 pub(crate) async fn bind_agent_skill(
     artifact_path: PathBuf,
@@ -30,102 +27,4 @@ pub(crate) async fn bind_agent_skill(
     );
 
     json_output(&bind_result_json(&artifact, operator, &result))
-}
-
-pub(crate) fn bind_result_json(
-    artifact: &nexus_sdk::types::TapPublishArtifact,
-    operator: sui::types::Address,
-    result: &BindAgentSkillResult,
-) -> serde_json::Value {
-    json!({
-        "standard_tap": true,
-        "function": "bind_agent_skill",
-        "digest": result.tx_digest,
-        "tx_checkpoint": result.tx_checkpoint,
-        "operator": operator,
-        "agent_id": result.agent_id,
-        "agent_object_id": result.agent_object.object_id(),
-        "agent_object_version": result.agent_object.version(),
-        "skill_id": result.skill_id,
-        "dag_id": artifact.dag_id,
-        "tap_package_id": artifact.tap_package_id,
-        "config_digest_hex": hex::encode(&result.config_digest),
-        "config_digest_input": result.config_digest_input,
-    })
-}
-
-#[cfg(test)]
-mod tests {
-    use {
-        super::*,
-        nexus_sdk::{
-            nexus::tap::BindAgentSkillResult,
-            types::{
-                InterfaceRevision,
-                TapConfigDigestInput,
-                TapPaymentPolicy,
-                TapPublishArtifact,
-                TapSchedulePolicy,
-                TapSkillConfig,
-                TapSkillRequirements,
-                TapVertexAuthorizationSchema,
-            },
-        },
-    };
-
-    fn fixture_artifact() -> TapPublishArtifact {
-        let config = TapSkillConfig {
-            name: "weather skill".to_string(),
-            tap_package_name: "weather_tap".to_string(),
-            dag_path: PathBuf::from("dag.json"),
-            tap_package_path: PathBuf::from("tap"),
-            requirements: TapSkillRequirements {
-                input_schema_commitment: vec![1],
-                workflow_commitment: vec![2],
-                metadata_commitment: vec![3],
-                payment_policy: TapPaymentPolicy::default(),
-                schedule_policy: TapSchedulePolicy::default(),
-                vertex_authorization_schema: TapVertexAuthorizationSchema::default(),
-            },
-            shared_objects: vec![],
-            interface_revision: InterfaceRevision(1),
-        };
-        TapPublishArtifact::from_config(
-            &config,
-            sui::types::Address::from_static("0xdd"),
-            sui::types::Address::from_static("0xee"),
-        )
-        .expect("artifact builds")
-    }
-
-    #[test]
-    fn bind_result_json_exposes_combined_evidence() {
-        let artifact = fixture_artifact();
-        let result = BindAgentSkillResult {
-            tx_digest: sui::types::Digest::from([7u8; 32]),
-            tx_checkpoint: 100,
-            agent_id: sui::types::Address::from_static("0xa1"),
-            agent_object: sui::types::ObjectReference::new(
-                sui::types::Address::from_static("0xa1"),
-                3,
-                sui::types::Digest::from([5u8; 32]),
-            ),
-            skill_id: 7,
-            config_digest: vec![9u8; 32],
-            config_digest_input: TapConfigDigestInput {
-                interface_revision: InterfaceRevision(1),
-                shared_objects: vec![],
-                requirements: artifact.requirements.clone(),
-            },
-        };
-        let json = bind_result_json(&artifact, sui::types::Address::from_static("0x2"), &result);
-        assert_eq!(json["function"], "bind_agent_skill");
-        assert_eq!(
-            json["agent_id"],
-            serde_json::json!(sui::types::Address::from_static("0xa1").to_string())
-        );
-        assert_eq!(json["skill_id"], serde_json::json!(7));
-        assert_eq!(json["config_digest_hex"].as_str().unwrap().len(), 64);
-        assert_eq!(json["tx_checkpoint"], serde_json::json!(100));
-    }
 }
