+++ cli/src/tap/tap_registry.rs
@@ -20,15 +20,3 @@ pub(crate) async fn show_registry() -> AnyResult<(), NexusCliError> {
 
     json_output(&registry_show_result_json(&registry))
 }
-
-pub(crate) fn registry_show_result_json(
-    registry: &nexus_sdk::types::TapRegistry,
-) -> serde_json::Value {
-    json!({
-        "id": registry.id,
-        "default_executor": registry.default_executor,
-        "agents": registry.agents,
-        "skills": registry.skills,
-        "endpoints": registry.endpoints,
-    })
-}
